Move theme to constants
